remove searchBarLayout story
