+++ src/stories/imdb/stories/index.stories.js
@@ -17,7 +17,3 @@ storiesOf('Live Demos|IMDB Search', module)
 storiesOf('Live Demos|IMDB Search', module)
   .addDecorator(ThemePicker('greyVest'))
   .addWithJSX('Dynamic Search Layout', require('./searchLayout').default)
-
-storiesOf('Live Demos|IMDB Search', module)
-  .addDecorator(ThemePicker('greyVest'))
-  .addWithJSX('Search Bar Layout', require('./searchBarLayout').default)
